Update OOBE
Updated Script with the Path for the local install.wim. This should prevent us from having to download Windows files everytime OSD runs. 
If the local files are not found or the search for the local file spits out an error then this should revert to downloading files from online, just as before. 
If this completely fails PLEASE REVERT TO THE PREVIOUS VERSION!!!